eng-dados/requirements.txt
unidecode
shapely
memory_profiler
line_profiler
kml2geojson
googlemaps
geopy
geopandas
xmltodict
pyocr
pendulum
Multi-Template-Matching
pandas_gbq
six
cligj
munch
UliEngineering
unidecode
googlemaps
